+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added normalization for invalid "yi-" and "wu-" syllable upon syllable separation.
 - Added equal sign (`=`) removal upon syllable separation.
+- Added support for merging "p" with the next token upon syllable separation.
 
 ### Fixed